On branch master  Changes to be committed: 	new file:   Album.cs 	new file:   Genero.cs 	modified:   Musica.cs 	modified:   Program.cs Final do Capitulo 4

+++ Musica.cs
@@ -1,7 +1,7 @@
 ﻿class Musica {
     public string Nome { get; set; }
     public string Artista { get; set; }
-    public float Duracao { get; set; }
+    public int Duracao { get; set; }
     public bool Disponivel { get; set; }
     public string DescricaoResumida => $"A música {Nome} pertence à banda {Artista}";
